requirements.txt
flask
requests
pycryptodome==3.10.1
Pyrebase4==4.7.1